[common_test] Don't abort test if opening of connection fails
When opening a connection, the connection process would link itself to
ct_util_server before calling the init callback, e.g. in ct_netconfc,
ct_telnet etc. If the init callback failed, then ct_util_server would
get the 'EXIT'. ct_util looks up the pid in the connection table, but
since the connection is not yet registered it is not found. ct_util
does not know which process it is and will thus die - aborting the
complete test run.

This commit moves the link(CtUtilServer) after the init callback, so a
crash in the init callback will not be detected by ct_util_server -
the caller process, however, will get a 'DOWN' message due to
monitoring and ct_gen_conn:start will give an error return.